adding a failing test
It looks like iorg--org-to-html does not actually leave any contents in
the simple.html file which it creates.  I'm not sure why this is, it may
be a bug with org-export-to-file, it could be a problem with our filter
which adds the button or it could be a problem with the way you are
calling this function.

To run the failing test evaluate the simple-test.el buffer, and the run
`ert', on my system every assertion but the last succeeds.